Drill unsigned long long through problem sets and call backs.

+++ lib/fsm.c
@@ -1912,7 +1912,7 @@ assert(fsm->lpath != NULL);
 	    rpmts ts = fsmGetTs(fsm);
 	    rpmfi fi = fsmGetFi(fsm);
 	    void * ptr;
-	    unsigned int archivePos = fdGetCpioPos(fsm->cfd);
+	    unsigned long long archivePos = fdGetCpioPos(fsm->cfd);
 	    if (archivePos > fi->archivePos) {
 		fi->archivePos = archivePos;
 		ptr = rpmtsNotify(ts, fi->te, RPMCALLBACK_INST_PROGRESS,
